refactor: feed ui를 여러 곳에서 사용 가능하도록 리펙토링
feed ui만 따로 모듈화
