+++ FE/src/components/common/InfoSection.tsx
@@ -2,6 +2,7 @@ import Link from "next/link";
 import FundingInfoTag from "../fundingDetail/Info/FundingInfoTag";
 import { FundingMember, PledgeFunding } from "@/Model/Funding";
 
+//member는 올란사람이다.
 export default function InfoSection({
   funding,
   member,
