--- FE/src/components/member/MemberHeader.tsx
@@ -1,8 +1,0 @@
-export default function MemberHeader() {
-  return (
-    <div className="flex items-center text-color121 text-[1.875rem] font-bold mt-11">
-      <h1 className="shrink-0 mr-[2.375rem]">마이페이지</h1>
-      <hr className="w-full" />
-    </div>
-  );
-}
